backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
pydantic==2.10.3
python-dotenv==1.0.1
google-api-python-client==2.155.0
google-auth-oauthlib==1.2.1
cryptography==44.0.0
httpx==0.28.1
